requirements.txt
# Core dependencies
python-dotenv>=0.19.0
requests>=2.26.0
pymongo>=4.0.0
pandas>=1.3.0
numpy>=1.21.0

# NLP
nltk>=3.6.0
textblob>=0.17.1
vaderSentiment>=3.3.2
spacy>=3.1.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0.tar.gz

# Web framework
streamlit>=1.0.0
plotly>=5.3.0

# Testing
pytest>=6.2.5
pytest-cov>=2.12.0

# Development
black>=21.9b0
flake8>=3.9.2
mypy>=0.910